+++ src/Symfony/Bundle/TwigBundle/DependencyInjection/TwigExtension.php
@@ -116,7 +116,7 @@ class TwigExtension extends Extension
         $container->getDefinition('twig.cache_warmer')->replaceArgument(2, $config['paths']);
         $container->getDefinition('twig.template_iterator')->replaceArgument(2, $config['paths']);
 
-        $bundleHierarchy = $this->getBundleHierarchy($container);
+        $bundleHierarchy = $this->getBundleHierarchy($container, $config);
 
         foreach ($bundleHierarchy as $name => $bundle) {
             $namespace = $this->normalizeBundleName($name);
@@ -133,6 +133,11 @@ class TwigExtension extends Extension
         }
 
         if (file_exists($dir = $container->getParameter('kernel.root_dir').'/Resources/views')) {
+            $twigFilesystemLoaderDefinition->addMethodCall('addPath', array($dir));
+        }
+        $container->addResource(new FileExistenceResource($dir));
+
+        if (file_exists($dir = $container->getParameterBag()->resolveValue($config['default_path']))) {
             $twigFilesystemLoaderDefinition->addMethodCall('addPath', array($dir));
         }
         $container->addResource(new FileExistenceResource($dir));
@@ -181,7 +186,7 @@ class TwigExtension extends Extension
         }
     }
 
-    private function getBundleHierarchy(ContainerBuilder $container)
+    private function getBundleHierarchy(ContainerBuilder $container, array $config)
     {
         $bundleHierarchy = array();
 
@@ -195,6 +200,11 @@ class TwigExtension extends Extension
             }
 
             if (file_exists($dir = $container->getParameter('kernel.root_dir').'/Resources/'.$name.'/views')) {
+                $bundleHierarchy[$name]['paths'][] = $dir;
+            }
+            $container->addResource(new FileExistenceResource($dir));
+
+            if (file_exists($dir = $container->getParameterBag()->resolveValue($config['default_path']).'/bundles/'.$name)) {
                 $bundleHierarchy[$name]['paths'][] = $dir;
             }
             $container->addResource(new FileExistenceResource($dir));