+++ src/Symfony/Bundle/TwigBundle/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * deprecated `Symfony\Bundle\TwigBundle\Command\DebugCommand`, use `Symfony\Bridge\Twig\Command\DebugCommand` instead
  * deprecated relying on the `ContainerAwareInterface` implementation for `Symfony\Bundle\TwigBundle\Command\LintCommand`
+ * added option to configure default path templates (via `default_path`)
 
 3.3.0
 -----